kmk: Merged in changes from GNU make 3.82. Previous GNU make base version was gnumake-2008-10-28-CVS.
git-svn-id: http://svn.netlabs.org/repos/kbuild/trunk@2591 8c291193-6614-0410-a08d-adc6e07170bb

+++ src/kmk/kmkbuiltin.c
@@ -119,7 +119,7 @@ int kmk_builtin_command(const char *pszCmd, char ***ppapszArgvToSpawn, pid_t *pP
         argv[argc] = malloc(cch + 1);
         if (!argv[argc])
         {
-            printf("kmk_builtin: out of memory. argc=%d len=%d\n", argc, pszEnd - pszCmd + 1);
+            printf("kmk_builtin: out of memory. argc=%d len=%d\n", argc, (int)(pszEnd - pszCmd + 1));
             break;
         }
         memcpy(argv[argc], pszCmd, cch);
